Update Guts
Duplicated aspects of the GUI to create a separate box to serve as a to-do list with the idea of having one area for upcoming events, and one for more mundane tasks that belong on a to-do list. The function calls still need to be updated, and in many cases new functions need to be created. Perhaps there is a way to reuse the old functions by passing different values to them? Or slightly modifying them somehow so their output is displayed or stored differently as necessary.