Configured a reminder pop-up box
Changed this code in lines 81-88. The function it calls returns a boolean as well as a string containing a message about the event if any. The boolean allows the condition to be triggered only if there is an event today. 
I've laid the infrastructure for an early reminder message as well, but I'll have to go back and dig around a bit before I make it work.