Update Guts
Just added a call to the remove_colors() function in the remove_priority() function. When an item is removed from the list while still prioritized this causes an error because it causes a shift in the indexes but doesn't alter the indexes saved in the priorities file. A way to solve this error might be to call the remove priority function in the same place that the remove function is called, but if the remove priority function is called first then it should help with the bug. However, this still leaves the problems of the altered indexes. Maybe a function that subtracts one from each index each time an item from the list is removed? But then that wouldn't work if for all items in the list that precede the item that is removed. I'll think on this later (probably tomorrow) when I fix it.